+++ server/native/gamestate/src/game.rs
@@ -1,6 +1,6 @@
 use rand::{thread_rng, Rng};
 use rustler::{NifStruct, NifUnitEnum};
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 
 use crate::board::{Board, Tile};
 use crate::character::{Character, Effect, Name};
@@ -221,8 +221,8 @@ impl GameState {
     pub fn move_with_joystick(
         self: &mut Self,
         player_id: u64,
-        x: f64,
-        y: f64,
+        x: f32,
+        y: f32,
     ) -> Result<(), String> {
         let player = Self::get_player_mut(&mut self.players, player_id)?;
         if matches!(player.status, Status::DEAD) {
@@ -345,7 +345,7 @@ impl GameState {
             let projectile = Projectile::new(
                 *next_projectile_id,
                 attacking_player.position,
-                JoystickValues::new(direction.x as f64 / 100f64, direction.y as f64 / 100f64),
+                JoystickValues::new(direction.x as f32 / 100f32, direction.y as f32 / 100f32),
                 14,
                 10,
                 attacking_player.id,
@@ -484,14 +484,14 @@ impl GameState {
         next_projectile_id: &mut u64,
     ) -> Result<(), String> {
         if direction.x != 0 || direction.y != 0 {
-            let angle = (direction.y as f64).atan2(direction.x as f64); // Calculates the angle in radians.
+            let angle = (direction.y as f32).atan2(direction.x as f32); // Calculates the angle in radians.
             let angle_positive = if angle < 0.0 {
                 (angle + 2.0 * PI).to_degrees() // Adjusts the angle if negative.
             } else {
                 angle.to_degrees()
             };
 
-            let angle_modifiers = [-20f64, -10f64, 0f64, 10f64, 20f64];
+            let angle_modifiers = [-20f32, -10f32, 0f32, 10f32, 20f32];
 
             for modifier in angle_modifiers {
                 let projectile = Projectile::new(
@@ -617,7 +617,7 @@ impl GameState {
             let projectile = Projectile::new(
                 *next_projectile_id,
                 attacking_player.position,
-                JoystickValues::new(direction.x as f64 / 100f64, direction.y as f64 / 100f64),
+                JoystickValues::new(direction.x as f32 / 100f32, direction.y as f32 / 100f32),
                 14,
                 10,
                 attacking_player.id,
@@ -977,8 +977,8 @@ fn distance_to_center(player: &Player, center: &Position) -> f64 {
 }
 
 // We might want to abstract this into a Vector2 type or something, whatever.
-fn normalize_vector(x: f64, y: f64) -> (f64, f64) {
-    let norm = f64::sqrt(x.powf(2.) + y.powf(2.));
+fn normalize_vector(x: f32, y: f32) -> (f32, f32) {
+    let norm = f32::sqrt(x.powf(2.) + y.powf(2.));
     (x / norm, y / norm)
 }
 
@@ -1003,8 +1003,8 @@ fn generate_new_position(
 pub fn new_entity_position(
     height: usize,
     width: usize,
-    direction_x: f64,
-    direction_y: f64,
+    direction_x: f32,
+    direction_y: f32,
     entity_position: Position,
     entity_speed: i64,
 ) -> Position {
@@ -1016,8 +1016,8 @@ pub fn new_entity_position(
         then round the values.
     */
     let (movement_direction_x, movement_direction_y) = normalize_vector(-direction_y, direction_x);
-    let movement_vector_x = movement_direction_x * (speed as f64);
-    let movement_vector_y = movement_direction_y * (speed as f64);
+    let movement_vector_x = movement_direction_x * (speed as f32);
+    let movement_vector_y = movement_direction_y * (speed as f32);
 
     let mut new_position_x = old_x as i64 + (movement_vector_x.round() as i64);
     let mut new_position_y = old_y as i64 + (movement_vector_y.round() as i64);
