+++ server/native/gamestate/src/lib.rs
@@ -152,8 +152,8 @@ fn new_round(game: GameState, players: Vec<Player>) -> GameState {
 fn move_with_joystick(
     game: GameState,
     player_id: u64,
-    x: f64,
-    y: f64,
+    x: f32,
+    y: f32,
 ) -> Result<GameState, String> {
     let mut game_2 = game;
     game_2.move_with_joystick(player_id, x, y)?;
